Add support for pie charts.
This commit adds the piechart() function, which creates a pie chart.

In the process of creating this function, in order to avoid code
duplication, I refactored parts of linegraph() out to several new
subroutines.  This includes the backdrop, title, subtitle, and legend,
all of which can now be simply called from any graph-creating
function.

The grid that linegraph() does will probably end up being refactored
out to a function as well, but since the pie chart doesn't use it,
that didn't happen for this commit.

I also added a simple README file, in .pod format.