Add support for bar graphs.
This commit adds the bargraph() function, and an example of how to use it.

In adding the bar graph function, a couple more things were refactored
out of linegraph(), into get_maxima() and grid(), and some additional
flexibility was added to legend() and rect().

The examples were renamed to reflect what they are examples of.  This
will become important as more examples accumulate.

Also, the readme document now has second-level headings for the specific
types of graphs.