grid(): extend horizontal grid lines slightly to the left of the axis.
The vertical grid lines were already extended slightly below the axis,
so that they are visible even if there is a graph element (such as a
bar graph bar) directly covering the line.  This commit does the same
thing for the horizontal grid lines, extending them slightly to the
left of the vertical axis, in case the area directly to the right of
the axis is completely covered.  The need for this was exposed by the
recent addition of area graphs, which can often cover that area.