Allow legendwidth to be explicitly specified.
In the long term, I want to default to automatically calculating an
appropriate legend width, taking into account the lengths of the
various labels, and probably even default to auto-wrapping any really
long ones as long as the list of them is short (so that vertical space
isn't a problem).

But that's more work, and this doesn't interfere with it in any way
but provides a workaround for now.